+++ test/Stormpath.Configuration.Test/DefaultConfigTestCases/JsonTestCase.cs
@@ -236,20 +236,16 @@ namespace Stormpath.Configuration.Test.DefaultConfigTestCases
     },
     ""social"": {
         ""facebook"": {
-            ""uri"": ""/callbacks/facebook"",
-            ""scope"": ""email""
-        },
-        ""github"": {
-            ""uri"": ""/callbacks/github"",
-            ""scope"": ""user:email""
+            ""displayName"": ""Facebook"",
+            ""scope"": ""openid""
         },
         ""google"": {
-            ""uri"": ""/callbacks/google"",
-            ""scope"": ""email profile""
+            ""displayName"": ""Google"",
+            ""scope"": ""openid""
         },
         ""linkedin"": {
-            ""uri"": ""/callbacks/linkedin"",
-            ""scope"": ""r_basicprofile r_emailaddress""
+            ""displayName"": ""LinkedIn"",
+            ""scope"": ""openid""
         }
     },
     ""me"": {