+++ test/Stormpath.Configuration.Test/ModifiedConfigTestCases/JsonTestCase.cs
@@ -179,19 +179,19 @@ namespace Stormpath.Configuration.Test.ModifiedConfigTestCases
     },
     ""social"": {
       ""facebook"": {
-        ""uri"": ""/callbackz/facebook"",
+        ""displayName"": ""Bookface"",
         ""scope"": ""email birthday""
       },
-      ""github"": {
-        ""uri"": ""/callbackz/github"",
+      ""microsoft"": {
+        ""displayName"": ""M$"",
         ""scope"": ""user:everything""
       },
       ""google"": {
-        ""uri"": ""/callbackz/google"",
+        ""displayName"": ""Elgoog"",
         ""scope"": ""email profile friends""
       },
       ""linkedin"": {
-        ""uri"": ""/callbackz/linkedin"",
+        ""displayName"": ""Also M$"",
         ""scope"": ""email interests""
       }
     },